# feat: Bundle Server dependencies