--- a/dotcom-rendering/Containerfile
+++ b/dotcom-rendering/Containerfile
@@ -1,11 +1,5 @@
 # This container is used in our CICD pipelines for running E2E and regression tests.
-FROM node:16.19.0-alpine AS dependencies
-
-COPY ./node_modules /opt/app/dotcom-rendering/node_modules
-COPY ./dotcom-rendering/node_modules /opt/app/dotcom-rendering/dotcom-rendering/node_modules
-
-# Seperate dependencies into their own layer to allow for better caching
-FROM dependencies
+FROM node:16.19.0-alpine
 
 WORKDIR /opt/app/dotcom-rendering/dotcom-rendering