--- a/dotcom-rendering/cloudformation.yml
+++ b/dotcom-rendering/cloudformation.yml
@@ -247,12 +247,12 @@ Resources:
           usermod -a -G frontend aws-kinesis-agent-user
           cd /home/dotcom-rendering
 
-          aws --region eu-west-1 s3 cp s3://aws-frontend-artifacts/frontend/${Stage}/${App}/dist/${App}.zip ./
+          aws --region eu-west-1 s3 cp s3://aws-frontend-artifacts/frontend/${Stage}/${App}/${App}.zip ./
           unzip -q ${App}.zip -d ${App}
 
           chown -R dotcom-rendering:frontend ${App}
 
-          cd ${App}/dotcom-rendering
+          cd ${App}
 
           export TERM=xterm-256color
           export NODE_ENV=production