… dotcom-rendering/cdk/lib/__snapshots__/dotcom-rendering.test.ts.snap
… dotcom-rendering/package.json
… yarn.lock