## What does this change?

 - Uses Webpack to bundle our Server dependencies which lets us tree-shake a lot of unused code in our dependencies away.
 - Replaces `html-minifier` with its fork `html-minifier-terser`. `html-minifier` does not play nicely with Webpack due to its dependency on UglifyJS, meanwhile `html-minifier-terser` uses terser which works fine in Webpack (I believe its also the library Webpack uses by default for minifying JS?)
## Why?

This allows users and actions to download and run our container images faster. It probably speeds up deployments a bit too, but likely only by a minimal amount, and it won't be noticed anyway due to how we wait 60 seconds for deployments to warm up anyway.

|       Type                   |   Before  | After | Reduction |
|------------------------------|-----------|-------|-----------|
| Container Image (compressed) | 361MB     | 56MB  |    85%    |
| rendering.zip                | 400MB     | 12MB  |    97%    |

## Build times (test on my Mac)

Running `time make build`

Before: 109.61s 102.28s 103.73s
After: 116.93s 122.94s 121.46s

Running `time make riffraff-bundle`

Before: 142.59s 139.38s 136.92s
After: 123.14s 126.19s 128.63s

Publishing image to GHCR

Before: 34s
After: 10s

The webpack build takes significantly longer with these changes, which is what we'd expect as it has to do a lot more work to build the server bundle now, **however** we make up all that time by having far less data to compress when creating the final Riffraff ZIP bundle.

## Impact on dev server

EDIT: Just realised the dev server uses the same config, I was hoping with getting away with only making these changes for the prod server. Need to investigate impact.

EDIT 2: Adds 1.5 sec to dev server startup, should be able to have it revert to the old behaviour in dev mode though.