@AshCorr 2023-04-03
I believe Express and Log4js are trying to dynamically import dependencies and webpack doesn't know how to deal with it. For example:

```
function doesAThing() {
   var something = import("another thing");
}
```

In expresses case its trying to see if there are any template rendering libraries and load them, since we don't use that functionality of Express we're fine.

Log4js is a little different though, its looks like its trying to load an appender library, this is something that we might need for our ELK logging to work correctly, need to investigate further.