@OllysCoding 2023-04-03
> @AshCorr Log4js is a little different though, it's looks like its trying to load an appender library, this is something that we might need for our ELK logging to work correctly, need to investigate further.

Can you somehow manually tell webpack to bundle this library in the webpack server config, then the dynamic import would (hopefully) just work? Not sure exactly what you would need to do in the webpack config to tell it to bundle that appender library but I'm sure there's a way ...