@AshCorr 2023-04-03
> Can you somehow manually tell webpack to bundle this library in the webpack server config, then the dynamic import would (hopefully) just work? Not sure exactly what you would need to do in the webpack config to tell it to bundle that appender library but I'm sure there's a way ...

It looks like we're only using the `file` and `console` appenders from Log4js which are considered as "Core appenders" and always imported, so I think we're actually fine. Still, I'd love to try and figure out if theres a way we can get rid of the warning without just adding an ignore for it.